hw/mappers: reorganize files
